MAT-405: Fully separate (charity) campaign and metacampaign types
Should be type-level only changes, but will help us see what's required
as we implement serving campaigns and metacampaigns separately from
matchbot
