+++ src/app/campaign-details/campaign-details.component.spec.ts
@@ -78,9 +78,7 @@ describe('CampaignDetailsComponent', () => {
       summary: 'Test campaign description',
       title: 'Test Campaign!',
       updates: [],
-      usesSharedFunds: false,
       alternativeFundUse: 'Some information about what happens if funds are not used',
-      campaignCount: undefined,
       championOptInStatement: undefined,
       championRef: undefined,
       hidden: false,
