+++ src/app/campaign.service.spec.ts
@@ -61,9 +61,7 @@ describe('CampaignService', () => {
       summary: 'Some long summary',
       title: 'Some title',
       updates: [],
-      usesSharedFunds: false,
       alternativeFundUse: 'Some information about what happens if funds are not used',
-      campaignCount: undefined,
       championOptInStatement: undefined,
       championRef: undefined,
       hidden: false,
