+++ src/app/donation-start/donation-start-form/donation-start-form.component.spec.ts
@@ -130,9 +130,7 @@ describe('DonationStartForm', () => {
       summary: 'Some long summary',
       title: 'Some title',
       updates: [],
-      usesSharedFunds: false,
       alternativeFundUse: 'Some information about what happens if funds are not used',
-      campaignCount: undefined,
       championOptInStatement: undefined,
       championRef: undefined,
       hidden: false,
